renamed source config dao to service and service to management service (#1052)
