+++ src/main/java/org/opensearch/securityanalytics/threatIntel/sacommons/TIFSourceConfigManagementService.java
@@ -1,7 +1,7 @@
 package org.opensearch.securityanalytics.threatIntel.sacommons;
 
 import org.opensearch.core.action.ActionListener;
-public abstract class TIFSourceConfigService {
+public abstract class TIFSourceConfigManagementService {
     IndexTIFSourceConfigResponse indexTIFConfig(IndexTIFSourceConfigRequest request, ActionListener <IndexTIFSourceConfigResponse> listener){
         return null;
     }
